segmented start connection segmented setup connection
